+++ packages/referral/src/affiliate/sections/barChart.tsx
@@ -1,18 +1,18 @@
 
-import { Select, cn, format, subDays } from "@orderly.network/react";
+import { cn, format, subDays } from "@orderly.network/react";
 import { FC, useContext, useMemo, useState } from "react";
 import { ColmunChart, InitialBarStyle, InitialXAxis, InitialYAxis, emptyDataSource, emptyDataSourceYAxis } from "../../components/barChart";
 import { ReferralContext } from "../../hooks/referralContext";
 import { RefFilterMenu } from "../../components/refFilterMenu";
-import { useReferalRebateSummary } from "../../hooks/useReferalRebateSummary";
 import { generateData } from "../../utils/utils";
+import { useReferralRebateSummary } from "@orderly.network/hooks";
 
 type ChartDataType = "Commission" | "Referral vol.";
 
 export const BarChart: FC<{ className?: string }> = (props) => {
     const [filterType, setFiltetType] = useState<ChartDataType>("Commission");
 
-    const [ rebateSummary] = useReferalRebateSummary({
+    const [ rebateSummary] = useReferralRebateSummary({
         startDate: format(subDays(new Date(), 15), 'yyyy-MM-dd'),
         endDate: format(subDays(new Date(), 1), 'yyyy-MM-dd'),
     });